CORE-53745 Moved injection of request sending methods up to index.js.
